Move DeviceHardwareRepository, FirmwareReleaseRepository, and RadioConfigRepository to :core:data
